Minor improvements to meson and examples (#8)
* meson: Use subdirs to structure the project

[Why]
In the existing form it was not possible to build out of tree, since
the path to the source file was wrong. Also installing the library
was not possible.

[How]
Use the subdir feature of meson to split up tasks. Then fix things
where they occur.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>

* plotter: Move all function in namespace

[Why]
It is good practice to move global functions to an anonymous namspace
so they are limited to the compilation unit and the compiler can tell
us if they are unused.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>

* plotter: Make examples more accessible

[Why]
It is quite tedious to change + compile + execute the file for every
example we wanna test.

[How]
Accept arguments and use them to select the corresponding sample.
We keep it as simple as possible, so no proper error handling.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>

---------

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>